Add GTFS download table

+++ repositories/impl/stop.py
@@ -13,7 +13,7 @@ class StopRepository:
     
     database: Database
     
-    def create(self, context: Context, row):
+    def create(self, download_id: int, context: Context, row):
         '''Inserts a new stop into the database'''
         stop_id = row['stop_id']
         number = row['stop_code']
@@ -22,8 +22,7 @@ class StopRepository:
         self.database.insert(
             table='stop',
             values={
-                'agency_id': context.agency_id,
-                'system_id': context.system_id,
+                'download_id': download_id,
                 'stop_id': stop_id,
                 'number': number,
                 'name': row['stop_name'],
@@ -38,22 +37,27 @@ class StopRepository:
         '''Returns the stop with the given context and stop ID'''
         stops = self.database.select(
             table='stop',
-            columns=[
-                'agency_id',
-                'system_id',
-                'stop_id',
-                'number',
-                'name',
-                'lat',
-                'lon',
-                'parent_id',
-                'type'
-            ],
+            columns={
+                'download.agency_id': 'agency_id',
+                'download.system_id': 'system_id',
+                'stop.stop_id': 'stop_id',
+                'stop.number': 'number',
+                'stop.name': 'name',
+                'stop.lat': 'lat',
+                'stop.lon': 'lon',
+                'stop.parent_id': 'parent_id',
+                'stop.type': 'type'
+            },
+            joins={
+                'download': {
+                    'download.download_id': 'stop.download_id'
+                }
+            },
             filters={
-                'agency_id': context.agency_id,
-                'system_id': context.system_id,
-                'stop_id': stop_id,
-                'number': number
+                'download.agency_id': context.agency_id,
+                'download.system_id': context.system_id,
+                'stop.stop_id': stop_id,
+                'stop.number': number
             },
             limit=1,
             initializer=Stop.from_db
@@ -66,34 +70,39 @@ class StopRepository:
     def find_all(self, context: Context, stop_number: str | None = None, limit: int | None = None, lat: float | None = None, lon: float | None = None, size: float = 0.01, parent_id: str | None = None, type: StopType | None = None) -> list[Stop]:
         '''Returns all stops that match the given context'''
         filters = {
-            'agency_id': context.agency_id,
-            'system_id': context.system_id,
-            'number': stop_number,
-            'parent_id': parent_id,
-            'type': type.value if type else None
+            'download.agency_id': context.agency_id,
+            'download.system_id': context.system_id,
+            'stop.number': stop_number,
+            'stop.parent_id': parent_id,
+            'stop.type': type.value if type else None
         }
         if (lat is not None and lon is not None):
-            filters['lat'] = {
+            filters['stop.lat'] = {
                 '>=': lat,
                 '<=': lat + size
             }
-            filters['lon'] = {
+            filters['stop.lon'] = {
                 '>=': lon,
                 '<=': lon + size
             }
         return self.database.select(
             table='stop',
-            columns=[
-                'agency_id',
-                'system_id',
-                'stop_id',
-                'number',
-                'name',
-                'lat',
-                'lon',
-                'parent_id',
-                'type'
-            ],
+            columns={
+                'download.agency_id': 'agency_id',
+                'download.system_id': 'system_id',
+                'stop.stop_id': 'stop_id',
+                'stop.number': 'number',
+                'stop.name': 'name',
+                'stop.lat': 'lat',
+                'stop.lon': 'lon',
+                'stop.parent_id': 'parent_id',
+                'stop.type': 'type'
+            },
+            joins={
+                'download': {
+                    'download.download_id': 'stop.download_id'
+                }
+            },
             filters=filters,
             limit=limit,
             initializer=Stop.from_db
@@ -102,25 +111,30 @@ class StopRepository:
     def find_matches(self, context: Context, query: str) -> list[Match]:
         stops = self.database.select(
             table='stop',
-            columns=[
-                'agency_id',
-                'system_id',
-                'stop_id',
-                'number',
-                'name',
-                'lat',
-                'lon',
-                'parent_id',
-                'type'
-            ],
+            columns={
+                'download.agency_id': 'agency_id',
+                'download.system_id': 'system_id',
+                'stop.stop_id': 'stop_id',
+                'stop.number': 'number',
+                'stop.name': 'name',
+                'stop.lat': 'lat',
+                'stop.lon': 'lon',
+                'stop.parent_id': 'parent_id',
+                'stop.type': 'type'
+            },
+            joins={
+                'download': {
+                    'download.download_id': 'stop.download_id'
+                }
+            },
             filters={
-                'agency_id': context.agency_id,
-                'system_id': context.system_id,
+                'download.agency_id': context.agency_id,
+                'download.system_id': context.system_id,
                 'OR': {
-                    'number': {
+                    'stop.number': {
                         'LIKE': f'%{query}%'
                     },
-                    'name': {
+                    'stop.name': {
                         'LIKE': f'%{query}%'
                     }
                 }
@@ -134,18 +148,23 @@ class StopRepository:
         areas = self.database.select(
             table='stop',
             columns={
-                'MIN(lat)': 'min_lat',
-                'MAX(lat)': 'max_lat',
-                'MIN(lon)': 'min_lon',
-                'MAX(lon)': 'max_lon'
+                'MIN(stop.lat)': 'min_lat',
+                'MAX(stop.lat)': 'max_lat',
+                'MIN(stop.lon)': 'min_lon',
+                'MAX(stop.lon)': 'max_lon'
+            },
+            joins={
+                'download': {
+                    'download.download_id': 'stop.download_id'
+                }
             },
             filters={
-                'agency_id': context.agency_id,
-                'system_id': context.system_id,
-                'lat': {
+                'download.agency_id': context.agency_id,
+                'download.system_id': context.system_id,
+                'stop.lat': {
                     '!=': 0
                 },
-                'lon': {
+                'stop.lon': {
                     '!=': 0
                 }
             },
@@ -158,10 +177,30 @@ class StopRepository:
     
     def delete_all(self, context: Context):
         '''Deletes all stops for the given context from the database'''
+        download_ids = self.database.select(
+            table='stop',
+            columns={
+                'stop.download_id': 'download_id'
+            },
+            distinct=True,
+            joins={
+                'download': {
+                    'download.download_id': 'stop.download_id'
+                }
+            },
+            filters={
+                'download.agency_id': context.agency_id,
+                'download.system_id': context.system_id
+            },
+            initializer=lambda r: r['download_id']
+        )
+        if not download_ids:
+            return
+        if len(download_ids) == 1:
+            download_ids = download_ids[0]
         self.database.delete(
             table='stop',
             filters={
-                'agency_id': context.agency_id,
-                'system_id': context.system_id
+                'download_id': download_ids
             }
         )
